Add Item class and make Participant a child class of Item
